rename emulation to syscall_mock
