+++ src/syscall_mock/test_result.rs
